// frontend/src/api/generated/model/dealCandidateResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Upbid API
 * ## 인증 방법

**로컬 환경**: 상단 `개발용 인증 > [로컬 전용] 테스트 로그인`을 먼저 호출하세요. 테스트 유저가 자동 생성되고 SESSION 쿠키가 발급됩니다.

**그 외 환경**: 카카오 로그인 후 발급된 SESSION 쿠키가 자동으로 포함됩니다.
 * OpenAPI spec version: v1
 */
import type { DealCandidateResponseDtoDealStatus } from './dealCandidateResponseDtoDealStatus';

export interface DealCandidateResponseDto {
  dealCandidateId?: number;
  candidateRank?: number;
  nickname?: string;
  bidAmount?: number;
  dealStatus?: DealCandidateResponseDtoDealStatus;
  phoneNumber?: string;
  isMe?: boolean;
}